pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = ["Eduard Schumacher <107512513+mijamoto911@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.30.0"
alembic = "^1.14.1"
greenlet = "^3.1.1"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
pydantic = {extras = ["email"], version = "^2.10.5"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
python-multipart = "^0.0.20"
bcrypt = "^4.2.1"
fastapi-mail = "^1.4.2"
slowapi = "^0.1.9"
cloudinary = "^1.42.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"
aiosqlite = "^0.21.0"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"
python-dotenv = "^1.0.1"



[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"



